BREAKING CHANGE: isolate bitreader usage to rbsp module
This will make a bitreader->bitstream-io transition easier.

I also got rid of the infrequently-used read_{ue,se} methods in favor of
read_{ue,se}_named. The latter are newer and provide better diagnostics.

Because this is a breaking change, I bumped the version number and
added a change log.