new, friendlier push API
No more need to deal with a user context, Box<RefCell<>>, or separate
traits for the various handlers. In the simplest case, just a closure
will do.

This is essentially performance-neutral by itself. It allows RBSP
parsing to be lazy though, and after the next commit that will pretty
significantly speed up the case where slice NALs are processed in a
single push.