touch-ups to `rbsp` interface
These improve suitability for parsing H.265 as well as H.264.

The changes to `ByteReader` construction are because H.265 has
two-byte NAL headers, so automatically skipping one byte is super
confusing. Arguably it was confusing even with H.264. Make the skip
explicit.

Additionally, `BitRead::read_to` is nice for loading stuff directly
into primitives such as `[u8; 11]`. That's the representation that
makes the most sense for the H.265 profile (as defined in section 7.3.3,
`profile_tier_level`, `if( profilePresentFlag )` block). When
constructing a `HEVCDecoderConfigurationRecord`, we need that full
11-byte span; when constructing the RFC 6381 codec string, we need
a 6-byte span. So parsing the whole thing into individual fields means
reconstructing these later.

And finally, I renamed methods to more closely match `bitstream-io`,
which is what `h264-reader` is using internally now and is
overwhelmingly the commonly used crate for bitstream parsing in general.