fix test post-#76
d95a87b fixed a parsing case that was erroneously consuming some extra
data. The test input (taken from a fuzz test, according to the comment)
appears to have been depending on the previous incorrect behavior to
exercise the desired logic. Come up with a new input that produces the
desired response.